requirements.txt
# PolyScalping - Requirements
# Python 3.10+

# CORE FRAMEWORK (Flask)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# ASYNC & HTTP
aiohttp>=3.9.1
httpx>=0.26.0
requests>=2.31.0

# AI / LLM APIs
openai>=1.10.0
anthropic>=0.18.0

# FINANCIAL DATA SOURCES
yfinance>=0.2.36
scipy>=1.11.4
numpy>=1.26.3

# BLOCKCHAIN / WEB3
web3>=6.15.1
eth-account>=0.11.0
py-clob-client>=0.15.0

# DATA PROCESSING
pandas>=2.1.4
python-dateutil>=2.8.2

# PAYMENTS
stripe>=7.0.0

# ENVIRONMENT & SECURITY
python-dotenv>=1.0.0
cryptography>=42.0.0

# TELEGRAM BOT
python-telegram-bot>=21.0

# IMAGE GENERATION
Pillow>=10.0.0

# UTILITIES
python-slugify>=8.0.1
pytz>=2023.3
